+++ src/Bullet.jsx
@@ -4,32 +4,32 @@ import classNames from 'classnames';
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Bullet = (props) => {
-  // bullet for paginator
-  const {children, defaultStyle} = props;
+import SliderPaginator from './SliderPaginator';
 
-  const style = {
-    display: 'block',
-    height: '20px',
-    width: '20px'
-  }
-
+/**
+ * Bullet
+ *  
+ * @param {any} props 
+ * @returns 
+ */
+const Bullet = ({index, children, defaultStyle, active, onClick}) => {
+  
   const handleClick = (event) => {
     event.preventDefault();
-    props.onClick(props.index, true);
+    onClick(index, true);
   };
 
   const classes = classNames(
     'viewport-slider-paginator-bullet',
-    { 'is-active': props.active }
-  )
+    { 'is-active': active }
+  );
 
   return (
     <a href="#"
       className={classes}
       onClick={handleClick}
       
-      {...defaultStyle ? {style} : {}}
+      {...defaultStyle ? {style:SliderPaginator.bulletDefaultStyle} : {}}
     >
       {children}
     </a>
@@ -37,19 +37,24 @@ const Bullet = (props) => {
 
 };
 
+Bullet.defaultProps = {
+  defaultStyle: true
+}
+
 Bullet.propTypes = {
+  /** Flag to set bullet active */
   active: PropTypes.bool,
+  /** Index to item panel */
   index: PropTypes.number.isRequired,
+  /** Click handler */
   onClick: PropTypes.func,
+  /** Element to set as children */
   children: PropTypes.oneOfType([
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node
   ]),
+  /** Flag to indicate default style should be used */
   defaultStyle: PropTypes.bool
 };
-
-Bullet.defaultProps = {
-  defaultStyle: true
-}
 
 export default Bullet;